Feat: add & update genre

+++ server/controllers/movieController.js
@@ -1,11 +1,4 @@
-const {
-  User,
-  Movie,
-  Cast,
-  Genre,
-  sequelize,
-  Sequelize: { op },
-} = require("../models");
+const { User, Movie, Cast, Genre, sequelize } = require("../models");
 
 class movieController {
   static async readAllMovies(req, res, next) {
@@ -214,7 +207,15 @@ class movieController {
   static async deleteMovieById(req, res, next) {
     try {
       const { id } = req.params;
-      const data = await Movie.destroy({
+      let data = await Movie.findByPk(req.params.id);
+      if (!data) {
+        throw {
+          code: 404,
+          msg: "Movie not found",
+        };
+      }
+
+      await Movie.destroy({
         where: {
           id,
         },
@@ -233,6 +234,44 @@ class movieController {
       });
     } catch (error) {
       next(error);
+    }
+  }
+
+  static async createNewGenre(req, res, next) {
+    try {
+      const { name } = req.body;
+      await Genre.create({ name });
+      res.status(201).json({
+        statusCode: 201,
+        message: "Genre is created successfully",
+      });
+    } catch (err) {
+      next(err);
+    }
+  }
+
+  static async updateGenre(req, res, next) {
+    try {
+      const { name } = req.body;
+      const { id } = req.params;
+      await Genre.update({ name }, { where: { id } });
+      res.status(201).json({
+        statusCode: 201,
+        message: "Genre is updated successfully",
+      });
+    } catch (err) {
+      next(err);
+    }
+  }
+
+  static async deleteTheGenre(req, res, next) {
+    try {
+      let data = await Genre.findByPk(req.params.id);
+      if (!data) throw { name: "Not Found" };
+      await Genre.destroy({ where: { id: req.params.id } });
+      res.status(200).json({ message: ` ${data.name} success to delete` });
+    } catch (err) {
+      next(err);
     }
   }
 }
